+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Nuvlabox status offline distributor - bugfix when an update of a NuvlaBox
   status fail, process was aborted #245
 - NuvlaEdge connector - improve and fix update_nuvlabox_engine #244
+- NuvlaEdge connector - add support for nuvlaedge org and main branch 
 
 ## [3.1.1] - 2022-08-25